+++ tests/apps/slack_integration/slash_command_test.py
@@ -9,17 +9,93 @@ from busy_beaver.apps.slack_integration.slash_command import (
     relink_github,
     upcoming_events,
 )
-from busy_beaver.models import GitHubSummaryUser
+from busy_beaver.models import GitHubSummaryUser, SlackUser
 
 pytest_plugins = ("tests._utilities.fixtures.slack",)
+
+
+##################
+# Integration Tests
+###################
+@pytest.mark.integration
+def test_slack_command_valid_command(
+    client, session, factory, create_slack_headers, generate_slash_command_request
+):
+    installation = factory.SlackInstallation()
+    data = generate_slash_command_request("help", team_id=installation.workspace_id)
+    headers = create_slack_headers(100_000_000, data, is_json_data=False)
+
+    response = client.post("/slack/slash-command", headers=headers, data=data)
+
+    assert response.status_code == 200
+    assert "/busybeaver help" in response.json["text"].lower()
+
+
+@pytest.mark.integration
+def test_slack_command_invalid_command(
+    client, session, factory, create_slack_headers, generate_slash_command_request
+):
+    installation = factory.SlackInstallation()
+    data = generate_slash_command_request(
+        "non-existent", team_id=installation.workspace_id
+    )
+    headers = create_slack_headers(100_000_000, data, is_json_data=False)
+
+    response = client.post("/slack/slash-command", headers=headers, data=data)
+
+    assert response.status_code == 200
+    assert "command not found" in response.json["text"].lower()
+
+
+@pytest.mark.integration
+def test_slack_command_empty_command(
+    client, session, factory, create_slack_headers, generate_slash_command_request
+):
+    installation = factory.SlackInstallation()
+    data = generate_slash_command_request(command="", team_id=installation.workspace_id)
+    headers = create_slack_headers(100_000_000, data, is_json_data=False)
+
+    response = client.post("/slack/slash-command", headers=headers, data=data)
+
+    assert response.status_code == 200
+    assert "/busybeaver help" in response.json["text"].lower()
+
+
+@pytest.mark.integration
+def test_slack_command_creates_user_record_in_database(
+    client, session, factory, create_slack_headers, generate_slash_command_request
+):
+    # Arrange
+    installation = factory.SlackInstallation()
+    data = generate_slash_command_request("help", team_id=installation.workspace_id)
+    headers = create_slack_headers(100_000_000, data, is_json_data=False)
+
+    # Act
+    client.post("/slack/slash-command", headers=headers, data=data)
+
+    # Assert
+    users = SlackUser.query.all()
+    assert len(users) == 1
+
+    user = users[0]
+    assert user.slack_id == data["user_id"]
+    assert user.installation.workspace_id == data["team_id"]
 
 
 ########################
 # Miscellaneous Commands
 ########################
 @pytest.mark.unit
-def test_command_help(generate_slash_command_request):
-    data = generate_slash_command_request("help")
+def test_command_help(session, factory, generate_slash_command_request):
+    github_summary_config = factory.GitHubSummaryConfiguration()
+    install = github_summary_config.slack_installation
+    new_user = "new_user"
+    slack_user = factory.SlackUser(slack_id=new_user, installation=install)
+    data = generate_slash_command_request(
+        "help", user_id=new_user, team_id=install.workspace_id
+    )
+    data["user"] = slack_user
+    data["installation"] = install
 
     result = display_help_text(**data)
 
@@ -32,12 +108,13 @@ def test_command_not_found(generate_slash_command_request):
 
     result = command_not_found(**data)
 
+    assert "not found" in result["text"]
     assert "/busybeaver help" in result["text"]
 
 
-##########################################
-# Associate GitHub account with Slack user
-##########################################
+#################
+# Upcoming Events
+#################
 @pytest.mark.unit
 def test_connect_command_new_user(session, factory, generate_slash_command_request):
     github_summary_config = factory.GitHubSummaryConfiguration()
